A first iteration of Chader, polling file changes at each frame

+++ src/app/chader/GenericDrawer.h
@@ -23,7 +23,7 @@ struct GenericDrawer
     {
         if (mVertexCount)
         {
-            assert(mVertices.size() == mVertexCount);
+            assert(static_cast<std::size_t>(mVertices.size()) == mVertexCount);
         }
         else
         {
